ADD: ButtonAnimationModel, UPDATE: FreePlayMain
ButtonAnimationModel used to animate the Cajon if struck on the right order, but for now the input of the user is just a simple 'click' on the screen.

FreePlayMain is updated due to the new added animation into the project.

This is a new branch which could be my workbench for now.